refactor(runtimed): split peer session bootstrap
