+++ crates/runtimed/src/notebook_sync_server/mod.rs
@@ -58,6 +58,7 @@ mod peer;
 mod peer_pool_sync;
 mod peer_presence;
 mod peer_runtime_sync;
+mod peer_session;
 mod peer_writer;
 mod persist;
 mod project_context;
